chore: move scan related models under scan directory
