+++ deltacat/storage/model/scan/push_down.py
