+++ deltacat/storage/model/scan/scan_task.py
